raspberry_pi_src/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "raspbery_pi_src"

dependencies=[
    "pyserial==3.5",
    "requests==2.32.5",
    "RPi.GPIO==0.7.1",
]

[project.scripts]
main = "raspbery_pi_src.main:__main__"
led_breather = "raspbery_pi_src.control_sys.LEDStripController:main"
system_test = "raspbery_pi_src.system_test:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
    "raspbery_pi_src" = ["config/*.json"]